requirements.txt
pyodbc>=5.0
sqlalchemy>=2.0
PyQt6>=6.6
PyMuPDF>=1.24
pyyaml>=6.0
pytest-qt>=4.2
fastapi>=0.110
uvicorn>=0.23
